+++ src/Vankosoft/CmsBundle/Model/FileManagerFile.php
@@ -5,9 +5,6 @@ class FileManagerFile extends File implements FileManagerFileInterface
     /** @var FileManagerInterface */
     protected $filemanager;
     
-    /** @var string */
-    protected $originalName;
-    
     public function getFilemanager(): FileManagerInterface
     {
         //return $this->filemanager;
@@ -18,18 +15,6 @@ class FileManagerFile extends File implements FileManagerFileInterface
     {
         $this->filemanager  = $filemanager;
         $this->setOwner( $filemanager );
-        
-        return $this;
-    }
-    
-    public function getOriginalName(): string
-    {        
-         return $this->originalName;
-    }
-    
-    public function setOriginalName( string $originalName ): self
-    {
-        $this->originalName = $originalName;
         
         return $this;
     }